move non-bridge method to helpers file
